// F1UI/F1-UI-Angular/src/app/api/model/driverTeam.ts
/**
 * F1Website
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DriverTeam { 
    driverId?: number;
    teamId?: number;
    dateFrom?: string;
    dateTo?: string | null;
    raceNumber?: number | null;
}